unicapture: extract converter module
